Linux: Fix build break in r40786 due to file not ending with a newline. (Don't know why the trybot didn't catch that ...) Also fix a GCC signedness warning when compiling the domain whitelist code.
TEST=built on Linux
BUG=none

Review URL: http://codereview.chromium.org/669206

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40792 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 39278018206ee0a46d00af8ed8f7b0ae28036d0e

+++ o3d/plugin/cross/whitelist.cc
@@ -144,7 +144,7 @@ static bool IsDomainWhitelisted(const std::string &in_url) {
       host[i] = tolower(host[i]);
     }
 
-    for (int i = 0; i < arraysize(kDomainWhitelist); ++i) {
+    for (size_t i = 0; i < arraysize(kDomainWhitelist); ++i) {
       size_t pos = host.rfind(kDomainWhitelist[i]);
       if (pos != std::string::npos &&
           ((pos + strlen(kDomainWhitelist[i]) == host.length())))
